// pslpwcl/src/app/modules/pslpapi/model/linguaDich.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { GradoConLingua } from './gradoConLingua';
import { LinguaBlp } from './linguaBlp';
import { ModApprLingua } from './modApprLingua';

export interface LinguaDich { 
    codGradoConLingua: GradoConLingua;
    codLingua: LinguaBlp;
    codUserAggiorn?: string;
    codUserInserim?: string;
    dAggiorn?: Date;
    dInserim?: Date;
    flgCertificazioneAcquisita: string;
    id?: number;
    idCandidatura: number;
    idModApprLingua?: ModApprLingua;
    tipoCertificazione?: string;
}